[Refactor] 채팅 및 칭호 관련 오류 수정

+++ BookBridge/BookBridge/ViewModel/Chat/ChatMessageViewModel.swift
@@ -306,6 +306,27 @@ extension ChatMessageViewModel {
         
         completion()
     }
+    
+    // 나갔다가 다시 보내는 경우(상대방이 chatId가 있음)
+    func handleNoChatRoom(uid: String, partnerId: String, chatRoomListId: String, completion: () -> ()) {
+        let timestamp = Date()
+        
+        let query1 = FirebaseManager.shared.firestore.collection("User").document(uid).collection("chatRoomList").document(chatRoomListId)
+        
+        query1.setData([
+            "date": timestamp,
+            "id": chatRoomListId,
+            "isAlarm": true,
+            "newCount": 0,
+            "noticeBoardId": noticeBoardInfo.id,
+            "noticeBoardTitle": noticeBoardInfo.noticeBoardTitle,
+            "partnerId": partnerId,
+            "recentMessage": "",
+            "userId": uid
+        ])
+        
+        completion()
+    }
 }
 
 //MARK: 메시지 전송 (Image)
@@ -356,24 +377,26 @@ extension ChatMessageViewModel {
                         guard error == nil else { return }
                         print("Recipient saved message as well")
                     }
-                    
-                    partnerQuery.getDocument { documentSnapshot, error in
-                        guard error == nil else { return }
-                        guard let document = documentSnapshot else { return }
-                        
-                        
-                        partnerQuery.updateData([
-                            "date": timestamp,
-                            "newCount": (document.data()?["newCount"] as? Int ?? 0) + 1,
-                            "recentMessage": "사진"
-                        ])
-                    }
                 }
                 self.nestedGroupImage.leave()
             }
             
             self.nestedGroup.notify(queue: .main) {
-                self.selectedImages.removeAll()
+                let partnerQuery = FirebaseManager.shared.firestore.collection("User").document(partnerId).collection("chatRoomList").document(self.saveChatRoomId)
+                
+                partnerQuery.getDocument { documentSnapshot, error in
+                    guard error == nil else { return }
+                    guard let document = documentSnapshot else { return }
+                    
+                    
+                    partnerQuery.updateData([
+                        "date": timestamp,
+                        "newCount": (document.data()?["newCount"] as? Int ?? 0) + self.selectedImages.count,
+                        "recentMessage": "사진"
+                    ])
+                    
+                    self.selectedImages.removeAll()
+                }
             }
         }
     }
@@ -630,21 +653,64 @@ extension ChatMessageViewModel {
 
 //MARK: 채팅방 나가기
 extension ChatMessageViewModel {
-    func deleteChatRoom(uid: String, completion: @escaping() -> ()) {
+    func deleteChatRoom(uid: String, partnerId: String, completion: @escaping() -> ()) {
         firestoreListener?.remove()
         
-        let messageRemoveQuery = FirebaseManager.shared.firestore.collection("User").document(uid).collection("chatRoomList").document(saveChatRoomId).collection("messages")
-        
-        messageRemoveQuery.getDocuments { querySnapshot, error in
+        FirebaseManager.shared.firestore.collection("User").document(partnerId).collection("chatRoomList").whereField("id", isEqualTo: self.saveChatRoomId).getDocuments { querySnapshot, error in
             guard error == nil else { return }
             guard let documents = querySnapshot?.documents else { return }
             
-            for document in documents {
-                messageRemoveQuery.document(document.documentID).delete()
-            }
+            let messageRemoveQuery = FirebaseManager.shared.firestore.collection("User").document(uid).collection("chatRoomList").document(self.saveChatRoomId).collection("messages")
             
-            FirebaseManager.shared.firestore.collection("User").document(uid).collection("chatRoomList").document(self.saveChatRoomId).delete { _ in
-                completion()
+            //상대방도 채팅방이 없을 경우만 이미지 삭제
+            if documents.isEmpty {
+                messageRemoveQuery.getDocuments { querySnapshot, error in
+                    guard error == nil else { return }
+                    guard let documents = querySnapshot?.documents else { return }
+                    
+                    for document in documents {
+                        guard let imageURL = document.data()["imageURL"] as? String else { return }
+                        
+                        if imageURL != "" {
+                            self.deleteChatImages(imageURL: imageURL)
+                        }
+                        
+                        messageRemoveQuery.document(document.documentID).delete()
+                    }
+                    
+                    FirebaseManager.shared.firestore.collection("User").document(uid).collection("chatRoomList").document(self.saveChatRoomId).delete { _ in
+                        
+                        completion()
+                    }
+                }
+            } else {
+                messageRemoveQuery.getDocuments { querySnapshot, error in
+                    guard error == nil else { return }
+                    guard let documents = querySnapshot?.documents else { return }
+                    
+                    for document in documents {
+                        messageRemoveQuery.document(document.documentID).delete()
+                    }
+                    
+                    FirebaseManager.shared.firestore.collection("User").document(uid).collection("chatRoomList").document(self.saveChatRoomId).delete { _ in
+                        
+                        completion()
+                    }
+                }
+            }
+        }
+    }
+    
+    //채팅 이미지 삭제
+    func deleteChatImages(imageURL: String) {
+        //상대방도 채팅방이 없을 경우만 삭제
+        let ref = FirebaseManager.shared.storage.reference(forURL: imageURL)
+        
+        ref.delete { err in
+            if err == nil {
+                print("사진 삭제 성공")
+            } else {
+                print("사진 삭제 오류")
             }
         }
     }
